extract main_apps functionality to a main file from entry.py
